requirements.txt
amqp==1.4.9
anyjson==0.3.3
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.23
celery==3.1.23
coverage==4.1
decorator==4.0.10
defusedxml==0.4.1
Django==1.9.7
django-allauth==0.25.2
django-debug-toolbar==1.4
django-nose==1.4.3
djangorestframework==3.3.3
ecdsa==0.13
Fabric==1.11.1
ipdb==0.10.0
ipython==4.2.0
ipython-genutils==0.1.0
kombu==3.0.35
nose==1.3.7
oauthlib==1.1.2
paramiko==1.17.0
pexpect==4.1.0
pickleshare==0.7.2
ptyprocess==0.5.1
pycrypto==2.6.1
python3-openid==3.0.10
pytz==2016.4
requests==2.10.0
requests-oauthlib==0.6.1
simplegeneric==0.8.1
sqlparse==0.1.19
traitlets==4.2.1
uWSGI==2.0.13.1
